Correction orientation fusee (float -> double, calcul trajectoire)

+++ src/ship.c
@@ -40,9 +40,9 @@ void updateShips(Ship *ships, int count, Planet *planets, int planetCount) {
         // Pour le deplacement des fusees dans l'espace
         if (ships[i].state == MOVING_TO_TARGET || ships[i].state == RETURNING) {
             Planet *destination = (ships[i].state == MOVING_TO_TARGET) ? ships[i].target : ships[i].base;
-            float dx = destination->x - (ships[i].x + 64 / 2);
-            float dy = destination->y - (ships[i].y + 64 / 2);
-            float distance = sqrt(dx * dx + dy * dy);
+            double dx = destination->x - (ships[i].x + 64 / 2.);
+            double dy = destination->y - (ships[i].y + 64 / 2.);
+            double distance = sqrt(dx * dx + dy * dy);
 
             if (distance - ships[i].speed >= destination->radius) {
                 ships[i].x += (dx / distance) * ships[i].speed;
@@ -79,13 +79,13 @@ void renderShips(SDL_Renderer *renderer, SDL_Texture *spriteSheet, Ship *ships, 
         // Calcul de l'angle en radians puis conversion en degrés
         double target_x = ships[i].target->x;
         double target_y = ships[i].target->y;
-        angle = atan2(target_y - (ships[i].y + 64 / 2 * camera.scale), target_x - (ships[i].x + 64 / 2 * camera.scale));
+        angle = atan2(target_y - (ships[i].y + 64 / 2.f * camera.scale), target_x - (ships[i].x + 64 / 2.f * camera.scale));
         angle = angle * (180.0f / M_PI);  // Conversion radians → degrés
         angle += (ships[i].state == MOVING_TO_TARGET) ? 90.0f : -90.0f;  // Si la fusée ne va pas vers la cible, on l'inverse
 
         // Calcul des coordonnées à l'écran
-        float screenX = (ships[i].x - camera.rect.x - SCREEN_WIDTH / 2) * camera.scale + SCREEN_WIDTH / 2;
-        float screenY = (ships[i].y - camera.rect.y - SCREEN_HEIGHT / 2) * camera.scale + SCREEN_HEIGHT / 2;
+        float screenX = (ships[i].x - camera.rect.x - SCREEN_WIDTH / 2.f) * camera.scale + SCREEN_WIDTH / 2.f;
+        float screenY = (ships[i].y - camera.rect.y - SCREEN_HEIGHT / 2.f) * camera.scale + SCREEN_HEIGHT / 2.f;
 
         SDL_Rect srcRect = {ships[i].frameIndex * 64, 0, 64, 64}; // Frame actuelle sur le sprite sheet
         SDL_Rect destRect = {(int)screenX, (int)screenY, (int)(64 * camera.scale), (int)(64 * camera.scale)}; // Position et taille affichée
